Update main movies fetch
