+++ src/app/components/main/Main.tsx
@@ -1,16 +1,10 @@
-import type { MoviesResult } from '@/interfaces';
 import { MoviesCard } from './';
 
-interface Props {
-  popularMovies: MoviesResult[];
-  topRatedMovies: MoviesResult[];
-}
-
-export const Main = ({ popularMovies, topRatedMovies }: Props) => {
+export const Main = () => {
   return (
     <main className='bg-[#13114b] px-4 py-5 text-white sm:px-8 md:px-16'>
-      <MoviesCard movies={popularMovies} moviesTitle='Top Movies Today' />
-      <MoviesCard movies={topRatedMovies} moviesTitle='Top Rated Movies' />
+      <MoviesCard movieType='popular' moviesTitle='Top Movies Today' />
+      <MoviesCard movieType='top_rated' moviesTitle='Top Rated Movies' />
     </main>
   );
 };
